newFile

+++ app/src/main/java/com/example/mobileappasm/Donation.kt
@@ -5,11 +5,11 @@ class Donation (
     var cardNo: String? = null,
     var cardCvv: String? = null,
     var cardExp: String? = null,
-    var amount: Int,
+    var amount: Double,
     var date: String? = null,
     var time: String? = null,
     var childName: String? = null,
     var username: String? = null
 ) {
-    constructor() : this("", "","", "", 0, "", "", "", "")
+    constructor() : this("", "","", "", 0.0, "", "", "", "")
 }
